requirements.txt
basicsr>=1.4.2
basicsr-fixed>=1.3.6
facexlib>=0.2.5
realesrgan>=0.3.0
torch
torchvision
torchaudio
numpy
opencv-python
Pillow 